Bypass IllegalArgumentException for implicit mutable flags using FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT.
Starting with U, for apps that target SDK U or higher, creation of a PendingIntent with FLAG_MUTABLE and an implicit Intent within will throw an IllegalArgumentException for security reasons. To bypass this check, we have used FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT when creating a PendingIntent, for now. However, the recommendation is not to use this flag and make the Intent explicit or the PendingIntent immutable, thereby making the Intent safe.

Follow up bug: 295850359

Bug: 294493834
Test: Trivial Change
Change-Id: I99af8e1b4c542553b7d7f983e9d7ba6a4261e467